+++ src/main/java/com/jokim/sivillage/api/common/config/SecurityConfig.java
@@ -1,0 +1,67 @@
+package com.jokim.sivillage.api.common.config;
+
+//import com.jokim.sivillage.common.jwt.JwtAuthenticationFilter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
+
+import java.util.List;
+
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+public class SecurityConfig {
+
+    private final AuthenticationProvider authenticationProvider;
+//    private final JwtAuthenticationFilter jwtAuthenticationFilter;
+
+    @Bean
+    public CorsFilter corsFilter() {
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowCredentials(true);
+        config.addAllowedOriginPattern("*");
+        config.addAllowedHeader("*");
+        config.addAllowedMethod("*");
+        config.setExposedHeaders(List.of("Authorization"));
+        source.registerCorsConfiguration("/**", config);
+        return new CorsFilter(source);
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+       http
+               .csrf(AbstractHttpConfigurer::disable)
+               .authorizeHttpRequests(
+                       authorizeRequests -> authorizeRequests
+                               .requestMatchers(
+                                       "/api/v1/auth/**",
+                                       "/swagger-ui/**",
+                                       "/v3/api-docs/**",
+                                       "/api/v1/**",
+                                       "/error"
+                               )
+                               .permitAll()
+                               .anyRequest()
+                               .authenticated()
+               )
+               .sessionManagement(
+                          sessionManagement -> sessionManagement
+                                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+               )
+               .authenticationProvider(authenticationProvider)
+//               .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
+               .addFilter(corsFilter());
+       return http.build();
+    }
+}
