+++ src/main/java/com/jokim/sivillage/api/customer/application/CustomerService.java
@@ -1,0 +1,7 @@
+package com.jokim.sivillage.api.customer.application;
+
+import com.jokim.sivillage.api.customer.dto.CustomerSignUpDto;
+
+public interface CustomerService {
+    void signUp(CustomerSignUpDto customerSignUpDto);
+}
