+++ src/main/java/com/jokim/sivillage/api/customer/infrastructure/CustomerAdressRepository.java
@@ -1,0 +1,7 @@
+package com.jokim.sivillage.api.customer.infrastructure;
+
+import com.jokim.sivillage.api.customer.domain.Address;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerAdressRepository extends JpaRepository<Address, Long> {
+}
