+++ src/main/java/com/jokim/sivillage/api/customer/infrastructure/CustomerDefaultAddresRepository.java
@@ -1,0 +1,7 @@
+package com.jokim.sivillage.api.customer.infrastructure;
+
+import com.jokim.sivillage.api.customer.domain.DefaultAddress;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerDefaultAddresRepository extends JpaRepository<DefaultAddress, Long> {
+}
