+++ src/main/java/com/jokim/sivillage/api/customer/infrastructure/CustomerMarketingRepository.java
@@ -1,0 +1,7 @@
+package com.jokim.sivillage.api.customer.infrastructure;
+
+import com.jokim.sivillage.api.customer.domain.Marketing;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerMarketingRepository extends JpaRepository<Marketing,Long> {
+}
