+++ src/main/java/com/jokim/sivillage/api/customer/vo/CustomerResponseVo.java
@@ -1,0 +1,22 @@
+package com.jokim.sivillage.api.customer.vo;
+
+import com.jokim.sivillage.api.customer.domain.State;
+import lombok.*;
+
+import java.util.Date;
+
+@Getter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class CustomerResponseVo {
+    private Long id;
+    private String customerUuid;
+    private String email;
+    private String password;
+    private String name;
+    private Date birth;
+    private String phone;
+    private State state;
+}
