--- src/main/java/com/jokim/sivillage/common/config/ApplicationConfig.java
@@ -1,49 +1,0 @@
-package com.jokim.sivillage.common.config;
-
-
-import com.jokim.sivillage.customer.infrastructure.CustomerRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@RequiredArgsConstructor
-@Configuration
-public class ApplicationConfig {
-
-    private final CustomerRepository customerRepository;
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return email -> {
-            return customerRepository.findByEmail(email).orElseThrow(
-                    () -> new IllegalArgumentException("해당 이메일을 가진 회원이 없습니다.")
-            );
-        };
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider daoAuthenticationProvider = new DaoAuthenticationProvider();
-        daoAuthenticationProvider.setUserDetailsService(userDetailsService());
-        daoAuthenticationProvider.setPasswordEncoder(passwordEncoder());
-        return daoAuthenticationProvider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration)
-        throws Exception {
-        return authenticationConfiguration.getAuthenticationManager();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
