--- src/main/java/com/jokim/sivillage/common/entity/CommonResponseEntity.java
@@ -1,17 +1,0 @@
-package com.jokim.sivillage.common.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CommonResponseEntity<T> {
-
-    private HttpStatus status;
-    private String message;
-    private T data;
-
-}
