--- src/main/java/com/jokim/sivillage/customer/infrastructure/CustomerAdressRepository.java
@@ -1,7 +1,0 @@
-package com.jokim.sivillage.customer.infrastructure;
-
-import com.jokim.sivillage.customer.domain.Address;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CustomerAdressRepository extends JpaRepository<Address, Long> {
-}
