--- src/main/java/com/jokim/sivillage/customer/infrastructure/CustomerDefaultAddresRepository.java
@@ -1,7 +1,0 @@
-package com.jokim.sivillage.customer.infrastructure;
-
-import com.jokim.sivillage.customer.domain.DefaultAddress;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CustomerDefaultAddresRepository extends JpaRepository<DefaultAddress, Long> {
-}
