--- src/main/java/com/jokim/sivillage/customer/infrastructure/CustomerMarketingRepository.java
@@ -1,7 +1,0 @@
-package com.jokim.sivillage.customer.infrastructure;
-
-import com.jokim.sivillage.customer.domain.Marketing;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CustomerMarketingRepository extends JpaRepository<Marketing,Long> {
-}
