--- src/main/java/com/jokim/sivillage/customer/infrastructure/CustomerPolicyRepository.java
@@ -1,7 +1,0 @@
-package com.jokim.sivillage.customer.infrastructure;
-
-import com.jokim.sivillage.customer.domain.Policy;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CustomerPolicyRepository extends JpaRepository<Policy, Long> {
-}
